requirement.txt
bleak==0.20.2
matplotlib==3.8.4
opencv-python==4.9.0.80
pyttsx3==2.90
pywavelets==1.6.0
tensorflow==2.15.0
Pillow==10.2.0
